+++ databox/testing/helpers.ts
@@ -8,13 +8,17 @@ import * as net from 'net';
 import * as http2 from 'http2';
 import Core from '@ulixee/hero-core';
 import { CanceledPromiseError } from '@ulixee/commons/interfaces/IPendingWaitEvent';
-import Hero, { ConnectionToHeroCore, IHeroCreateOptions } from '@ulixee/hero';
+import { ConnectionToHeroCore } from '@ulixee/hero';
 import DataboxInternal from '@ulixee/databox-for-hero/lib/DataboxInternal';
 import IDataboxForHeroRunOptions from '@ulixee/databox-for-hero/interfaces/IDataboxForHeroRunOptions';
 import TransportBridge from '@ulixee/net/lib/TransportBridge';
 import { Helpers } from './index';
 
 export const needsClosing: { close: () => Promise<any> | void; onlyCloseOnFinal?: boolean }[] = [];
+
+export function onClose(closeFn: (() => Promise<any>) | (() => any), onlyCloseOnFinal = false) {
+  needsClosing.push({ close: closeFn, onlyCloseOnFinal });
+}
 
 export interface ITestKoaServer extends KoaRouter {
   close: () => void;