+++ playgrounds/databox-for-hero/index.mjs
@@ -1,3 +1,4 @@
+import { setupAutorunMjsHack } from '@ulixee/databox/lib/utils/Autorun.mjs';
 import cjsImport from './index.js';
 
 const { Observable, Runner, Extractor } = cjsImport;
@@ -5,3 +6,5 @@ const { Observable, Runner, Extractor } = cjsImport;
 export { Observable, Runner, Extractor };
 
 export default cjsImport.default;
+
+setupAutorunMjsHack(cjsImport.default);